refactor(cctv): pecah page.tsx monolitik ke hooks dan komponen terpisah
- _types.ts        → CCTVNode, Zona, BoundingBox, DialogState, helpers
- _hooks/useDialog → toast + confirm dialog state
- _hooks/useCctv   → CCTV data, CRUD, AI toggle, stream state
- _hooks/useZona   → Zona data, CRUD, sort
- _components/StreamImg      → img dengan cleanup koneksi MJPEG on unmount
- _components/CctvCard       → kartu kamera individual
- _components/CctvModal      → form tambah/edit kamera
- _components/ZonaCard       → kartu zona individual
- _components/ZonaModal      → form tambah/edit zona
- _components/ConfirmDialog  → dialog alert/confirm + toast
- page.tsx          → slim orchestrator (~130 baris)
